fix: resolve build break from undefined functions in provider_helpers_test.go
The test file pkg/agent/provider_helpers_test.go referenced functions
(estimateTokensFromText, estimateChatTokenUsage, buildPromptWithHistoryGeneric,
etc.) that live in pkg/agent/providers, not pkg/agent. This caused compile
errors blocking Release, Nightly Tests, and Weekly Release workflows.

Fix: move tests for providers-package functions into the correct test file
(pkg/agent/providers/provider_helpers_test.go) and keep only tests for
agent-package functions (isPrivateIP, newRestrictedAIProviderHTTPClient)
in pkg/agent/provider_helpers_test.go.
